+++ src/include/sfptpd_crny_helper.h
@@ -6,6 +6,21 @@
 
 
 /****************************************************************************
+ * Types
+ ****************************************************************************/
+
+/* operation for clock control script */
+enum chrony_clock_control_op {
+	CRNY_CTRL_OP_NOP,
+	CRNY_CTRL_OP_ENABLE,
+	CRNY_CTRL_OP_DISABLE,
+	CRNY_CTRL_OP_SAVE,
+	CRNY_CTRL_OP_RESTORE,
+	CRNY_CTRL_OP_RESTORENORESTART,
+};
+
+
+/****************************************************************************
  * Function Prototypes
  ****************************************************************************/
 
@@ -14,5 +29,6 @@ extern int sfptpd_crny_helper_connect(const char *client_path,
 				      int *sock_ret,
 				      const char **failing_step);
 
+extern int sfptpd_crny_helper_control(enum chrony_clock_control_op op);
 
 #endif /* _SFPTPD_CRNY_HELPER_H */